Added 2018 Data
Finally added 2018 data by re-running the fetch_api function in espn-data. This required making edits to allow the function to run box_scores method vs scoreboard method
 - Created a new pickle file to retain the old one in case of errors
 - Rebuilt database and CSV views to include 2018 data (no player game data from 2018)

With the addition of 2018 data, made edits to week_page.py to allow it to reference YEARS_WEEKS and build all of the 2018 pages. In the process, rebuilt every single page. All pages were affected due to adding more to the navbar

New team added: Jackson

Current direction is to rebuild SVG's manually to save on space, and enable more of them to appear in website